debug a minor error in updating progress and status of a course & assignment

+++ lib/models/Course.dart
@@ -5,7 +5,7 @@ class Course {
 
   String Name;
   int Credit;
-  int Mark;
+  double Mark;
   String id;
   double progress = 0;
   var weeks = 0;
